+++ gsa/src/web/pages/reports/emptyresultsreport.js
@@ -20,6 +20,8 @@ import 'core-js/fn/string/includes';
 
 import React from 'react';
 
+import styled from 'styled-components';
+
 import _ from 'gmp/locale';
 
 import {isDefined} from 'gmp/utils/identity';
@@ -35,6 +37,11 @@ import ReportPanel from './reportpanel';
 import FilterIcon from 'web/components/icon/filtericon';
 import EditIcon from 'web/components/icon/editicon';
 import DeleteIcon from 'web/components/icon/deleteicon';
+
+const FilterString = styled.span`
+  font-style: italic;
+  margin-left: 10px;
+`;
 
 const EmptyResultsReport = ({
   all,
@@ -57,7 +64,10 @@ const EmptyResultsReport = ({
             '{{all}} results.',
           {all},
         )}
-      />
+      >
+        {_('Current applied filter is: ')}
+        <FilterString>{filter.toFilterString()}</FilterString>
+      </InfoPanel>
 
       <Divider align={['start', 'stretch']} wrap>
         {!levels.includes('g') && (