+++ CHANGELOG.md
@@ -24,6 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Cleanup get_report function in gsad [#1263](https://github.com/greenbone/gsa/pull/1263)
 
 ### Fixed
+- Fix filter selection at report details page [#1314](https://github.com/greenbone/gsa/pull/1314)
+- Fix using default results filter at report details page [#1314](https://github.com/greenbone/gsa/pull/1314)
 - Fix loading filters at permission list page [#1306](https://github.com/greenbone/gsa/pull/1306)
 - Fix filter in Report Results view cannot be saved & Fix error for create filter with no available results [#1303](https://github.com/greenbone/gsa/pull/1303)
 - Fix creating permissions via the create multiple permissions dialog [#1302](https://github.com/greenbone/gsa/pull/1302)